+++ docs/src/live/config/autoFormatRules/autoformatConstants.ts
