--- docs/src/live/config/autoFormatRules/index.ts
@@ -1,19 +1,0 @@
-export { basicElements } from './basicElements';
-export { basicLists } from './basicLists';
-export { basicMarks } from './basicMarks';
-export { punctuation } from './punctuation';
-export { legal, legalHtml } from './legal';
-export { arrows } from './arrows';
-export { mathEquality } from './mathEquality';
-export { mathComparison } from './mathComparison';
-export {
-  mathOperations,
-  mathMultiplication,
-  mathDivision,
-} from './mathOperations';
-export { mathFraction } from './mathFraction';
-export {
-  mathSuperscriptNumbers,
-  mathSuperscriptSymbols,
-} from './mathSuperscript';
-export { mathSubscriptNumbers, mathSubscriptSymbols } from './mathSubscript';
