+++ docs/src/live/config/autoFormatRules/text/autoformatArrow.ts
@@ -1,41 +1,41 @@
 import { AutoformatRule, ELEMENT_DEFAULT } from '@udecode/plate';
-import { formatText } from './_utils';
+import { formatText } from '../autoformatUtils';
 
-export const punctuation: AutoformatRule[] = [
+export const autoformatArrow: AutoformatRule[] = [
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
     markup: '-',
-    trigger: '-',
-    insertTrigger: false,
-    triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '\u2014'),
-  },
-  {
-    type: ELEMENT_DEFAULT,
-    mode: 'block',
-    markup: '..',
-    trigger: '.',
-    insertTrigger: false,
-    triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '…'),
-  },
-  {
-    type: ELEMENT_DEFAULT,
-    mode: 'block',
-    markup: '>',
     trigger: '>',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '»'),
+    format: (editor) => formatText(editor, '→'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
     markup: '<',
-    trigger: '<',
+    trigger: '-',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '«'),
+    format: (editor) => formatText(editor, '←'),
+  },
+  {
+    type: ELEMENT_DEFAULT,
+    mode: 'block',
+    markup: '=',
+    trigger: '>',
+    insertTrigger: false,
+    triggerAtBlockStart: false,
+    format: (editor) => formatText(editor, '⇒'),
+  },
+  {
+    type: ELEMENT_DEFAULT,
+    mode: 'block',
+    markup: ['<', '≤'],
+    trigger: '=',
+    insertTrigger: false,
+    triggerAtBlockStart: false,
+    format: (editor) => formatText(editor, '⇐'),
   },
 ];
