+++ docs/src/live/config/autoFormatRules/text/math/autoformatMathEquality.ts
@@ -1,59 +1,50 @@
 import { AutoformatRule, ELEMENT_DEFAULT } from '@udecode/plate';
-import { formatText } from './_utils';
+import { formatText } from '../../autoformatUtils';
 
-export const mathComparison: AutoformatRule[] = [
+export const autoformatMathEquality: AutoformatRule[] = [
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
     markup: '!',
-    trigger: '>',
-    insertTrigger: false,
-    triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '≯'),
-  },
-  {
-    type: ELEMENT_DEFAULT,
-    mode: 'block',
-    markup: '!',
-    trigger: '<',
-    insertTrigger: false,
-    triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '≮'),
-  },
-  {
-    type: ELEMENT_DEFAULT,
-    mode: 'block',
-    markup: '>',
     trigger: '=',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '≥'),
+    format: (editor) => formatText(editor, '≠'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '<',
+    markup: '=',
     trigger: '=',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '≤'),
+    format: (editor) => formatText(editor, '≡'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '!>',
+    markup: ['!=', '≠'],
     trigger: '=',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '≱'),
+    format: (editor) => formatText(editor, '≢'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '!<',
+    markup: '~',
     trigger: '=',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '≰'),
+    format: (editor) => formatText(editor, '≈'),
+  },
+  {
+    type: ELEMENT_DEFAULT,
+    mode: 'block',
+    markup: '!~',
+    trigger: '=',
+    insertTrigger: false,
+    triggerAtBlockStart: false,
+    format: (editor) => formatText(editor, '≉'),
   },
 ];
