+++ docs/src/live/config/autoFormatRules/text/math/autoformatMathSubscript.ts
@@ -1,125 +1,116 @@
 import { AutoformatRule, ELEMENT_DEFAULT } from '@udecode/plate';
-import { formatText } from './_utils';
+import { formatText } from '../../autoformatUtils';
 
-export const mathSuperscriptNumbers: AutoformatRule[] = [
+export const autoformatMathSubscriptNumbers: AutoformatRule[] = [
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '0',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '⁰'),
+    format: (editor) => formatText(editor, '₀'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '1',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '¹'),
+    format: (editor) => formatText(editor, '₁'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '2',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '²'),
+    format: (editor) => formatText(editor, '₂'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '3',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '³'),
+    format: (editor) => formatText(editor, '₃'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '4',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '⁴'),
+    format: (editor) => formatText(editor, '₄'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '5',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '⁵'),
+    format: (editor) => formatText(editor, '₅'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '6',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '⁶'),
+    format: (editor) => formatText(editor, '₆'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '7',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '⁷'),
+    format: (editor) => formatText(editor, '₇'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '8',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '⁸'),
+    format: (editor) => formatText(editor, '₈'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '9',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '⁹'),
+    format: (editor) => formatText(editor, '₉'),
   },
 ];
 
-export const mathSuperscriptSymbols: AutoformatRule[] = [
+export const autoformatMathSubscriptSymbols: AutoformatRule[] = [
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
-    trigger: 'o',
-    insertTrigger: false,
-    triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '°'),
-  },
-  {
-    type: ELEMENT_DEFAULT,
-    mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '+',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '⁺'),
+    format: (editor) => formatText(editor, '₊'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '^',
+    markup: '~',
     trigger: '-',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '⁻'),
+    format: (editor) => formatText(editor, '₋'),
   },
 ];
