+++ docs/src/live/config/autoFormatRules/text/math/autoformatMathSuperscript.ts
@@ -1,116 +1,125 @@
 import { AutoformatRule, ELEMENT_DEFAULT } from '@udecode/plate';
-import { formatText } from './_utils';
+import { formatText } from '../../autoformatUtils';
 
-export const mathSubscriptNumbers: AutoformatRule[] = [
+export const autoformatMathSuperscriptNumbers: AutoformatRule[] = [
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '0',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₀'),
+    format: (editor) => formatText(editor, '⁰'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '1',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₁'),
+    format: (editor) => formatText(editor, '¹'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '2',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₂'),
+    format: (editor) => formatText(editor, '²'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '3',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₃'),
+    format: (editor) => formatText(editor, '³'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '4',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₄'),
+    format: (editor) => formatText(editor, '⁴'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '5',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₅'),
+    format: (editor) => formatText(editor, '⁵'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '6',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₆'),
+    format: (editor) => formatText(editor, '⁶'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '7',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₇'),
+    format: (editor) => formatText(editor, '⁷'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '8',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₈'),
+    format: (editor) => formatText(editor, '⁸'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
     trigger: '9',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₉'),
+    format: (editor) => formatText(editor, '⁹'),
   },
 ];
 
-export const mathSubscriptSymbols: AutoformatRule[] = [
+export const autoformatMathSuperscriptSymbols: AutoformatRule[] = [
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
-    trigger: '+',
+    markup: '^',
+    trigger: 'o',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₊'),
+    format: (editor) => formatText(editor, '°'),
   },
   {
     type: ELEMENT_DEFAULT,
     mode: 'block',
-    markup: '~',
+    markup: '^',
+    trigger: '+',
+    insertTrigger: false,
+    triggerAtBlockStart: false,
+    format: (editor) => formatText(editor, '⁺'),
+  },
+  {
+    type: ELEMENT_DEFAULT,
+    mode: 'block',
+    markup: '^',
     trigger: '-',
     insertTrigger: false,
     triggerAtBlockStart: false,
-    format: (editor) => formatText(editor, '₋'),
+    format: (editor) => formatText(editor, '⁻'),
   },
 ];
